move mockEncrypted
