+++ test/modules/command/dh/dh-replication-import-command.test.js
@@ -24,7 +24,7 @@ const logger = require('../../../../modules/logger');
 
 const testUtilities = require('../../test-utilities');
 
-const {encryptedJSON, publicKey} = require('../../../../mockEncrypted.js');
+const {encryptedJSON, publicKey} = require('../../test_data/mockEncrypted.js');
 
 describe('Checks DHReplicationImportCommand execute() logic', function () {
     this.timeout(5000);
