+++ test/modules/test_data/mockEncrypted.js
@@ -1,4 +1,4 @@
-const Encryption = require('./modules/Encryption');
+const Encryption = require('../../../modules/Encryption');
 
 const jsonRaw = '{"@graph":[{"@type":"otObject","@id":"urn:uuid:4522de5e-9bc2-4a75-b04a-996a6ab17ce6","identifiers":[{"@type":"uuid","@value":"urn:uuid:4522de5e-9bc2-4a75-b04a-996a6ab17ce6"}],"relations":[{"@type":"otRelation","direction":"direct","linkedObject":{"@id":"urn:epc:id:sgtin:38512345.24121452145214.7VFZB0V6G545FZAW9YT7"},"properties":{"relationType":"EPC"}},{"@type":"otRelation","direction":"direct","linkedObject":{"@id":"urn:epc:id:sgtin:38512345.24121452145214.S4A9HXWHAXGZANE92DZM"},"properties":{"relationType":"EPC"}},{"@type":"otRelation","direction":"direct","linkedObject":{"@id":"urn:epc:id:sgtin:38512345.24121452145214.P25P0ZV6KB26TEZB24F6"},"properties":{"relationType":"EPC"}},{"@type":"otRelation","direction":"direct","linkedObject":{"@id":"urn:epc:id:sgtin:38512345.24121452145214.DN015S9641V45CFBCX8N"},"properties":{"relationType":"EPC"}},{"@type":"otRelation","direction":"direct","linkedObject":{"@id":"urn:epc:id:sgtin:38512345.24121452145214.ER72M16Y5Y7CAFEW1E93"},"properties":{"relationType":"EPC"}},{"@type":"otRelation","direction":"direct","linkedObject":{"@id":"urn:epc:id:sgtin:38512345.24121452145214.N0014KNEDRE337VG421W"},"properties":{"relationType":"EPC"}},{"@type":"otRelation","direction":"direct","linkedObject":{"@id":"urn:epc:id:sgtin:38512345.24121452145214.224F46H7H4X55B9B113C"},"properties":{"relationType":"EPC"}},{"@type":"otRelation","direction":"direct","linkedObject":{"@id":"urn:epc:id:sgtin:38512345.24121452145214.YEHB4NWA5WTBGKDABYGW"},"properties":{"relationType":"EPC"}},{"@type":"otRelation","direction":"direct","linkedObject":{"@id":"urn:epc:id:sgtin:38512345.24121452145214.N5Y3SGB9C25P1126XVKD"},"properties":{"relationType":"EPC"}},{"@type":"otRelation","direction":"direct","linkedObject":{"@id":"urn:epc:id:sgtin:38512345.24121452145214.B6VPXEMV8T6NPHPSVWWP"},"properties":{"relationType":"EPC"}},{"@type":"otRelation","direction":"direct","linkedObject":{"@id":"urn:epc:id:sgln:38512345.3851234500005"},"properties":{"relationType":"SOURCE"}},{"@type":"otRelation","direction":"direct","linkedObject":{"@id":"urn:epc:id:sgln:38512345.3851234500005"},"properties":{"relationType":"BIZ_LOCATION"}},{"@type":"otRelation","direction":"direct","linkedObject":{"@id":"urn:epc:id:sgln:38512345.3851234500005"},"properties":{"relationType":"READ_POINT"}}],"properties":{"objectType":"TransactionEvent","___metadata":{"eventTime":null,"eventTimeZoneOffset":null,"bizTransactionList":{"bizTransaction":[{"_attributes":{"type":"urn:epcglobal:cbv:btt:pedigree"}}]},"epcList":{"epc":[null,null,null,null,null,null,null,null,null,null]},"action":null,"bizStep":null,"disposition":null,"readPoint":{"id":null},"bizLocation":{"id":null},"extension":{"sourceList":{"source":[{"_attributes":{"type":"urn:epcglobal:cbv:sdt:owning_party"}}]}},"example0:expiry":{"_attributes":{"xsi:type":"xsd:dateTime"}},"example1:lot":{"_attributes":{"xsi:type":"xsd:int"}}},"eventTime":"2016-08-17T00:00:00.000Z","eventTimeZoneOffset":"+02:00","bizTransactionList":{"bizTransaction":["urn:epcglobal:cbv:bt:38512345.3851234500005:209"]},"epcList":{"epc":["urn:epc:id:sgtin:38512345.24121452145214.7VFZB0V6G545FZAW9YT7","urn:epc:id:sgtin:38512345.24121452145214.S4A9HXWHAXGZANE92DZM","urn:epc:id:sgtin:38512345.24121452145214.P25P0ZV6KB26TEZB24F6","urn:epc:id:sgtin:38512345.24121452145214.DN015S9641V45CFBCX8N","urn:epc:id:sgtin:38512345.24121452145214.ER72M16Y5Y7CAFEW1E93","urn:epc:id:sgtin:38512345.24121452145214.N0014KNEDRE337VG421W","urn:epc:id:sgtin:38512345.24121452145214.224F46H7H4X55B9B113C","urn:epc:id:sgtin:38512345.24121452145214.YEHB4NWA5WTBGKDABYGW","urn:epc:id:sgtin:38512345.24121452145214.N5Y3SGB9C25P1126XVKD","urn:epc:id:sgtin:38512345.24121452145214.B6VPXEMV8T6NPHPSVWWP"]},"action":"ADD","bizStep":"urn:epcglobal:cbv:bizstep:shipping","disposition":"urn:epcglobal:cbv:disp:in_progress","readPoint":{"id":"urn:epc:id:sgln:38512345.3851234500005"},"bizLocation":{"id":"urn:epc:id:sgln:38512345.3851234500005"},"extension":{"sourceList":{"source":["urn:epc:id:sgln:38512345.3851234500005"]}},"example0:expiry":"2018-06-08T14:58:56.591Z","example1:lot":"15123"}},{"@id":"urn:uuid:63212f6a-7223-477c-87c7-199d3c3601f4","@type":"otConnector","connectionId":"urn:epcglobal:cbv:bt:38512345.3851234500005:209","relations":[{"@type":"otRelation","direction":"reverse","linkedObject":{"@id":"urn:uuid:4522de5e-9bc2-4a75-b04a-996a6ab17ce6"},"properties":{"relationType":"CONNECTOR_FOR"}}]}],"@id":"0x624011867fbcd6248bd773915016048987c90523a0f77cabbb210bb54c85af79","@type":"Dataset","datasetHeader":{"OTJSONVersion":"1.0","datasetCreationTimestamp":"2019-06-06T13:32:14.024Z","datasetTitle":"","datasetTags":[],"relatedDatasets":[],"validationSchemas":{"erc725-main":{"schemaType":"ethereum-725","networkId":"rinkeby"},"merkleRoot":{"schemaType":"merkle-root","networkId":"rinkeby"}},"dataIntegrity":{"proofs":[{"proofValue":"0x3d55efb001f24fc0ae2e80afa056dee5d38fa15c7def92e3f7f713622cc867fe","proofType":"merkleRootHash","validationSchema":"/schemas/merkleRoot"}]},"dataCreator":{"identifiers":[{"identifierType":"ERC725","validationSchema":"/schemas/erc725-main","identifierValue":"0x2029a07cc3dc96a82b3260ebc1fed1d7"}]},"transpilationInfo":{"transpilationInfo":{"transpilerType":"GS1-EPCIS","transpilerVersion":"1.0","sourceMetadata":{"created":"2019-06-06T13:32:11.398Z","modified":"2019-06-06T13:32:11.398Z","standard":"GS1-EPCIS","XMLversion":"1.0","encoding":"UTF-8"},"diff":{}},"diff":{"_declaration":{"_attributes":{"version":"1.0","encoding":"UTF-8","standalone":"yes"}},"_doctype":"project","epcis:EPCISDocument":{"_attributes":{"schemaVersion":"1.2","xmlns:xsd":"http://www.w3.org/2001/XMLSchema","xmlns:xsi":"http://www.w3.org/2001/XMLSchema-instance","xmlns:example0":"http://ns.example.com/epcis0","xmlns:example1":"http://ns.example.com/epcis1","xmlns:example2":"http://ns.example.com/epcis2","xmlns:example3":"http://ns.example.com/epcis3","xmlns:example":"http://ns.example.com/epcis","creationDate":"2013-06-04T14:59:02.099+02:00","xmlns:epcis":"urn:epcglobal:epcis:xsd:1"},"EPCISBody":{}}}}},"signature":{"value":"0x18269f9ff4daefe903aba693d5647771015e45577d8fd3aa06a4b59268125dd537fce3956d59b4616be7a4a4b9aec766bbdf74a8ad8942939267723b2181b56d1c","type":"ethereum-signature"}}';
