+++ test/modules/test_data/otjson-graph.js
@@ -1048,7 +1048,7 @@ const graph3 = {
             "properties": {"relationType": "CONNECTOR_FOR"}
         }]
     }],
-    "@id": "0x4588b47fe852d2e74b835c05a5a5d602df84699f9d39e72ecee6a1bdac519ac2",
+    "@id": "0x624011867fbcd6248bd773915016048987c90523a0f77cabbb210bb54c85af79",
     "@type": "Dataset",
     "datasetHeader": {
         "OTJSONVersion": "1.0",
